+++ lib/rubocop/cop/sensei_cops.rb
@@ -2,6 +2,7 @@
 
 require_relative 'lecture/explain_bang_bang'
 require_relative 'lecture/explain_numbered_parameters'
+require_relative 'lecture/explain_or_assign'
 require_relative 'lecture/explain_pattern_match_case_in'
 require_relative 'lecture/explain_pattern_match_in'
 require_relative 'lecture/explain_pattern_match_rocket'