+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.7 (2023-xx-xx)
 
 - Lecture/ExplainBangBang
+- Lecture/ExplainOrAssign
 
 ## 0.1.6 (2023-07-18)